reservation edition 2.0

+++ shared/core-utils/src/types/common.types.ts
@@ -1,4 +1,4 @@
-import { Static, TSchema, Type } from '@sinclair/typebox'
+import { Static, TSchema, t as Type } from 'elysia'
 
 export const OptionalType = <T extends TSchema>(schema: T) => Type.Union([
   schema,
@@ -20,12 +20,12 @@ export const UuidSchema = Type.String({ format: 'uuid' })
 export const TimestampSchema = Type.String({ format: 'date-time' })
 export const NonEmptyStringSchema = Type.String({ minLength: 1 })
 export const PositiveNumberSchema = Type.Number({ minimum: 0 })
-export const PositiveIntegerSchema = Type.Integer({ minimum: 0 })
+export const PositiveIntegerSchema = Type.Numeric({ minimum: 0 })
 
 // Pagination schema
 export const PaginationParamsSchema = Type.Object({
-  page: Type.Integer({ minimum: 1, default: 1 }),
-  limit: Type.Integer({ minimum: 1, maximum: 100, default: 10 }),
+  page: Type.Numeric({ minimum: 1, default: 1 }),
+  limit: Type.Numeric({ minimum: 1, maximum: 100, default: 10 }),
   sortBy: Type.Optional(Type.String()),
   sortOrder: Type.Optional(Type.Union([
     Type.Literal('asc'),
@@ -48,10 +48,10 @@ export const ApiResponseSchema = <T extends TSchema>(dataSchema: T) => Type.Obje
   error: Type.Optional(ApiErrorSchema),
   meta: Type.Optional(Type.Object({
     pagination: Type.Optional(Type.Object({
-      total: Type.Integer(),
-      page: Type.Integer(),
-      limit: Type.Integer(),
-      totalPages: Type.Integer()
+      total: Type.Numeric(),
+      page: Type.Numeric(),
+      limit: Type.Numeric(),
+      totalPages: Type.Numeric()
     }))
   }))
 })
